Move area accumulation to generic functions
